requirements.txt
PyQt5==5.15.4
python-whois==0.7.3
requests==2.25.1
selenium==3.141.0
bs4==0.0.1
pandas==1.3.2
lxml==4.6.3